finished player movement
